requirements.txt
fastapi
uvicorn
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
